+++ R/supercells.R
@@ -8,7 +8,7 @@
 #' You can use either `k` or `step`.
 #' It is also possible to provide a set of points (an `sf` object) as `k` together with the `step` value to create custom cluster centers.
 #' @param compactness A compactness value. Larger values cause supercells to be more compact/even (square).
-#' Set `compactness = "auto"` to enable adaptive compactness (SLIC0).
+#' Use [use_adaptive()] to enable adaptive compactness (SLIC0).
 #' A compactness value depends on the range of input cell values and selected distance measure.
 #' @param dist_fun A distance function. Currently implemented distance functions are `"euclidean"`, `"jsd"`, `"dtw"` (dynamic time warping), name of any distance function from the `philentropy` package (see [philentropy::getDistMethods()]; "log2" is used in this case), or any user defined function accepting two vectors and returning one value. Default: `"euclidean"`
 #' @param avg_fun An averaging function - how the values of the supercells' centers are calculated? The algorithm internally implements common functions `"mean"` and `"median"` (provided with quotation marks), but also accepts any fitting R function (e.g., `base::mean()` or `stats::median()`, provided as plain function name: `mean`). Default: `"mean"`. See details for more information.
@@ -17,7 +17,9 @@
 #' @param minarea Specifies the minimal size of a supercell (in cells). Only works when `clean = TRUE`.
 #' By default, when `clean = TRUE`, average area (A) is calculated based on the total number of cells divided by a number of supercells
 #' Next, the minimal size of a supercell equals to A/(2^2) (A is being right shifted)
-#' @param step The distance (number of cells) between initial supercells' centers. You can use either `k` or `step`.
+#' @param step Initial center spacing. You can use either `k` or `step`.
+#' Provide a plain numeric value for cell units, or use [use_meters()] for
+#' map-distance steps in meters (automatically converted to cells using raster resolution).
 #' @param transform Transformation to be performed on the input. By default, no transformation is performed. Currently available transformation is "to_LAB": first, the conversion from RGB to the LAB color space is applied, then the supercells algorithm is run, and afterward, a reverse transformation is performed on the obtained results. (This argument is experimental and may be removed in the future).
 #' @param metadata Logical. Controls whether metadata columns
 #' ("supercells", "x", "y") are included.
@@ -75,9 +77,6 @@ supercells = function(x, k, compactness, dist_fun = "euclidean", avg_fun = "mean
   } else {
     "values"
   }
-  if (iter == 0) {
-    clean = FALSE
-  }
 
   centers_arg = NULL
   if (!missing(k) && inherits(k, "sf")) {
@@ -85,13 +84,7 @@ supercells = function(x, k, compactness, dist_fun = "euclidean", avg_fun = "mean
     k = NULL
   }
 
-  if (!inherits(x, "SpatRaster")) {
-    if (inherits(x, "stars")) {
-      x = terra::rast(x)
-    } else{
-      stop("The SpatRaster class is expected as an input", call. = FALSE)
-    }
-  }
+  x = .sc_util_prep_raster(x)
 
   trans = .supercells_transform_to_lab(x, transform)
   x = trans$x
@@ -105,7 +98,6 @@ supercells = function(x, k, compactness, dist_fun = "euclidean", avg_fun = "mean
     iter = iter,
     outcomes = outcomes,
     chunks = chunks,
-    iter_diagnostics = FALSE,
     verbose = verbose
   )
   if (!missing(step)) {
@@ -121,11 +113,8 @@ supercells = function(x, k, compactness, dist_fun = "euclidean", avg_fun = "mean
     args$centers = centers_arg
   }
 
-  if (iter == 0) {
-    slic_sf = do.call(sc_slic_points, args)
-  } else {
-    slic_sf = do.call(sc_slic, args)
-  }
+  slic_runner = if (iter == 0) sc_slic_points else sc_slic
+  slic_sf = do.call(slic_runner, args)
   if (isTRUE(trans$did_transform)) {
     names_x = names(x)
     slic_sf = .supercells_transform_from_lab(slic_sf, names_x)